Complete TimdexRecord model
Why these changes are being introduced:
There are fields in the TIMDEX data model that weren't yet added to this
TimdexRecord model because they haven't been required for the transforms
yet. This commit completes the model so it's ready for upcoming
transform work.

How this addresses that need:
* Adds missing fields call_numbers, contents, holdings, links,
  literary_form, numbering, physical_description, and
  publication_frequency to TimdexRecord model.
* Updates tests and conftest to reflect model changes.
* Also update .gitignore to include .python-version in source control
  and exclude vscode files.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/RDI-106