Update sentry and logging configuration
Why these changes are being introduced:

Our terraform template requires that each env variable be present in
all envs, even if we don't need it in a particular env (e.g. SENTRY_DSN
is not needed in dev). Env variables are now given a string value of
"None" in env(s) where they are not needed.

In addition, while working on this a logging bug was addressed in which
logging was configured on the cli module logger instead of the root
logger and thus was not propagating to all modules.

In resolving both issues it made sense to refactor these configuration
steps into their own functions in a separate config module.

How this addresses that need:
* Creates a new config module with functions to configure logging and
  sentry, and moves that logic from the cli module to these functions.
* Updates sentry configuration to check for a string value of "None" in
  addition to a None object value for the SENTRY_DSN env variable.
* Adds unit tests for new config module and update cli tests to match
  new expected outputs.

Additional changes:
* Minor update to conftest.py configuration for all tests.
* Two minor logging additions:
  * One to add a timer to the transform command and log the total
    transform time at the end of the process, useful as a benchmark for
    large/long transform jobs.
  * One to log a status update after every 1000 records transformed,
    useful to monitor progress during large/long transform jobs.
* Update Makefile to match standard section formatting across repos.
* Update Github Actions to use our shared CI workflow.
* Update coverage reporting to use coverage library instead of
  pytest-cov to match our shared CI workflow.
* Code/documentation cleanup across a few files.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/RDI-160
* https://mitlibraries.atlassian.net/browse/RDI-148