Refactor source transforms
Why these changes are being introduced:
There was a lot of code and functionality duplication across each of the
source transform classes. In addition, an increasingly long if/elif
chain was determining source transform instantiation directly in the cli
command. We'd like to remove the duplication and separate the class
instantiation concern from the cli module.

How this addresses that need:
* Adds a base Transformer class with initialization, iteration, and next
  method that is common to all source transformers.
* Makes source transformers subclasses of the base Transformer and
  removes their now unnecessary code.
* Moves transformer selection to the config module and determines it at
  runtime by the static configuration of each source.
* Bases the cli source option on the configured sources to ensure that
  definition of possible source transforms is located in only one place.
* Adds/updates tests to reflect all changes, including cleaning up
  fixtures and test names.
* Removes some no-longer-needed tests to reduce duplication.
* Adds an optional env variable to adjust the transform status update
  interval.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/RDI-165