Handle missing title field and refactor sources
Why these changes are being introduced:

ALthough title is a required field in TIMDEX, it is not present in some
source records. Rather than stopping the application with an error
when there is no title, we need to log it, add default title text, and
continue processing records.

In addition, after the previous source transform refactor all of the
required TIMDEX fields were still being populated in each source
transform with a lot of duplicated code across the sources. We want to
reduce the duplicate code and make it clear which methods must be
overridden in each subclass and which ones may not be overridden. Since
the title field population was part of this duplicated required field
code block, it made sense to do the refactor now so the title change
would only need to be made in one place.

How this addresses that need:

* Makes the base Transformer class a metaclass with several abstract
  methods to be implemented by each subclass.
* Moves all of the logic handling required TIMDEX fields into the base
  Transformer class, calling individual methods overridden in the
  subclasses for logic specific to each source.
* Adds a method in the base Transformer class to generate a valid title
  field value from a list of titles extracted from the source record,
  logging irregularities and inserting default text when no title is
  present.
* Adds override methods in all source subclasses as needed and refactors
  logic to populate and return optional field values for each source.
* Moves tests for logic now in the transformer class to the transformer
  test module.
* Updates tests and fixtures as needed to reflect code changes.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/RDI-187