Update subfield logic and tests across sources
Why these changes are being introduced:
Some fields that have subelements use the string value of those
subelements to create part of the TIMDEX field, but didn't check for the
presence of the string value in the subfield, resulting in an error.
Similarly, some fields that use the attribute presence to determine
logic weren't accounting for attributes that are present but empty, also
resulting in errors.

How this addresses that need:
* Adds tests and test fixtures to check for all variations of attributes
  and subfields (including not present and blank) when those are used to
  determine logic or create part of a TIMDEX field.
* Updates source transform logic to make sure new tests pass.
* Simplifies some of the xml parsing code.
* Updates logic to remove IsIdenticalTo from relatedItems in Datacite
  transform.
* Updates logic to include any extra main titles in alternate_titles
  across sources.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/RDI-233