Update WHOAS and Zenodo content type filters
Why these changes are being introduced:

Some adjustments need to be made to these content type filters per
stakeholder review after ingest with the initial filters. In addition to
updating the types we want to filter, a request was also made to not
ingest records without a content type in the source record from the
WHOAS source, instead of setting the default value of "Not specified".

How this addresses that need:
* Updates Zenodo content type filter to select records based on valid
  content types instead of invalid ones, and adds to valid types per
  stakeholder request.
* Updates dspace_dim transform to include a get_content_types method
  since the method needs to be different for the base transform vs.
  WHOAS
* Adds some types to the WHOAS invalid content types list, including
  "no content type in source record", and subclasses the
  get_content_type method to set content type to that value if there are
  no content types provided in the source record.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/RDI-243
* https://mitlibraries.atlassian.net/browse/RDI-246