Skip deleted records during transform
Why these changes are being introduced:
As we are not yet handling deleted records, we need to skip them during
transform so they don't cause errors in the transform process. In a
future commit we will replace this with actual deleted record handling.

How this addresses that need:
* Adds a method in the Transformer class to check for deleted record
  status.
* Updates the transformer.transform method to return None (thus skipping
  the record) if the record status is deleted.
* Updates tests and fixtures to reflect changes.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-126