Fix file open bug for s3
Why these changes are being introduced:
Due to current weirdness with the Alma records, the deleted records
output file was opened in append mode to ensure we didn't overwrite
deleted records that may show up in the wrong export file. However,
smart_open doesn't implement append mode in its open() function with S3
files so this needs to revert to write mode.

How this addresses that need:
* Changes the file open mode for helpers.write_deleted_records_to_file()
  to write.

Side effects of this change:
Deleted records during Alma export may not be 100% accurate until the
current Alma record export issue is fixed. The metadata team is working
on it and we will just have to do a full re-ingest once that's done to
ensure we have accurate data in TIMDEX.