Re-map MARC portfolio items to holdings and links
Why these changes are being introduced:
Our initial mapping for e-portfolio items (data enrichment field 986)
didn't include all the possible subfields where the URL could exist. In
addition, we weren't adding these to links, only holdings, which was
potentially confusing since they are links to the items.

How this addresses that need:
* Adds a new static method in the Marc class to get a single subfield
  value string if present in the datafield. This is more clear than
  using the method that is intended to retrieve multiple values, in
  cases like this where there will not be multiple values.
* Refactors the 986 field mapping to use the new static method to
  retrieve a single value from each single non-repeatable subfield.
* Adds additional subfield mappings for location value to the holdings
  mapping from the MARC 986 field.
* If an electronic holding has a location value, also add it as a Link
  (location value must be present because URL is required in the Link
  field).
* Adds/updates tests and fixtures to reflect changes.

Side effects of this change:
Electronic item links will be duplicated in the links and holdings
fields for some links. We decided this was ok for now.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-184