Change Springshare link and identifier parsing
Why these changes are being introduced:
After some analysis of transformed Springshare records, it was discovered that some timdex_record_ids were undesirable / poorly formed.  For a subset of records, the dc:identifier value was a PHP URL pattern, that while a functional link, did not result in a good timdex record identifier.

Previously, the source record id and the source base url were tightly coupled in the transformer class, and could not be overridden by transformations.  This meant identifiers for the record must be concate-able with the base url.  In this situation, this was not possible for all records in the set.

How this addresses that need:
  * added two new methods in the base transformer class, get_source_link() and get_timdex_record_id()
  * update springshare transform to utilize the get_source_link() to generate a valid URL
  * update springshare transform to fallback on the more widely used pattern of using part of the OAI identifier as the timdex record id

Side effects of this change:
Two new OPTIONAL methods are available for transformers, get_source_link() and get_timdex_record_id().

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-232