Update aspace identifiers logic
Why these changes are being introduced:

It was discovered that some ASpace records after transformation contained multiple identifiers, of the same type,
which was throwing errors in Bento.  It looks as though ASpace records may have changed slightly and are
exposing a new string for the <unitid> element, which is the same as the OAI record identifier.

How this addresses that need:

When iterating through <unitid> elements, by skipping those with attribute type=aspace_uri, we can
skip adding that particular <unitid> as a TIMDEX record identifier.

Side effects of this change:

Believed to be none.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/TIMX-234