Use @normal attribute for ASpace EAD dates
Why these changes are being introduced:
We receieved a Sentry error that a record had an invalid date range.  As described in the linked ticket,
it became apparent this was related to a combination of date parsing and validation, and how it varies
between Transmogrifier and OpenSearch environments.

It has been noted that the EAD unitdate field @normal attribute might contain a more normalized
form of the date that could help with these kind of errors or ambiguity.

How this addresses that need:
The date parsing for ASpace EADs has been updated to use the @normal attribute, which provides a
cleaner and more predictable string to work with for splitting as a date range.

Side effects of this change:
There are 60 ASpace EADs that do not have a @normal attribute in their archdesc.unitdate element.
For these, it's possible they will lose a date value if they had been parsed previously.  But it's
believed this will not be the case, given their element text values were all also somewhat unusual.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-235
* https://mitlibraries.atlassian.net/browse/TIMX-92