Allow Aardvark download links if scalar value
Why these changes are being introduced:

While technically valid for an Aardvark file to have multiple download links as part
of dct_references_s, when creating a TIMDEX record, we cannot know which link to include
as the sole and primary download link.  Single, scalar values are by far the most common and
will be used, but if an array of values is encountered, adding a Download link is skipped.

How this addresses that need:
* Ensure that value coming from dct_references_s for URI 'http://schema.org/downloadUrl' is a scalar value

Side effects of this change:
* Slight decrease in download links for transforming OGM records

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-109