Validate Date.value strings for Aardvark transformations
Why these changes are being introduced:

It was determined after some end-to-end testing that a large number of Aardvark records, when transformed,
were setting dates in the TIMDEX record that Opensearch could not parse.  This was unlike other transformations
where the date strings are validated.

NOTE: it is anticipated that we will return to date validation, possibly to extend it beyond just validation
and into extraction as well (e.g. pulling valid year '1933' from date string '1933?').

How this addresses that need:
* A try/except block added to field method _range_dates to warn and continue
* All non DateRange values are validated in field method get_dates; if fail validation, they are dropped

Side effects of this change:
* Massive increase in success rate for Opensearch indexing for gisogm source
* Decrease in date strings present in final records that are not valid without manipulation

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-109