MITAardvark get source link from source metadata
Why these changes are being introduced:

Unlike other Transmog sources, which dynamically create a source link for the record
based on a base URL and some identifier, MITAardvark records are expected to have a fully
formed URL already present in the metadata.  This is true for both MIT and OGM records, where
MIT records point to geo TIMDEX UI, and OGM records point to the holding institution.

How this addresses that need:
* Updates get_source_link() method to use data from metadata
* Sets config.base_url to string 'None' for both gismit and gisogm

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-109