Heavy rework of format/datatype ADR for GIS records
Why these changes are being introduced:

After multiple discussions about the various options outlined in the first version,
it was decided to take another pass at the ADR context and decision.  This reflects
some discussions to emphasize the data model as an evolving one, in which this
decision may feed into future decisions, while hopefully solving an immediate
issue for a TIMDEX source.

How this addresses that need:
* Reworks ADR 0002

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-200