Update post-transform methods to handle None values
Why these changes are being introduced:
* When a field method returns "None", the resulting "fields" dictionary will
set a key-value pair, where key = <field_name> and value = None. This will result
in the .setdefault() method being ignored during the post-transform methods
as it will only set the value in the "fields" dictionary to an empty list if
the key is not present.

How this addresses that need:
* Check if field methods that are updated by post-transform methods are
already set to None. If None, set to an empty list.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-217
* https://mitlibraries.atlassian.net/browse/GDT-210