Address liteary form control field bug
Why these changes are being introduced:

There is at least one MARC record that comes through
where control field 008 is not 34 characters, and therefore
was throwing an error when we attempt to grab the 33rd character.

How this addresses that need:
* If control field 008 is < 34 chars, return None for literary form

Side effects of this change:
* Transmogrifier runs for large swaths of files, e.g. full runs, will
not catastrophically fail for this single bug.
* This might suggest another layer of error handling to avoid full
run failure for a single record, will create at ticket for this.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-355