Add run_timestamp CLI argument
Why these changes are being introduced:

As outlined in TIMX-509, we need the ability for multiple Transmogrifier
runs to share the same run_timestamp.  Previously, the run_timestamp
was minted by the TIMDEXDataset.write() operation as part of the TDA
library.  With that now accepting an explicit run_timestamp for each
records, and Transmogrifier providing those records, we just needed
a way to pass a run_timestamp via the Transmogrifier CLI that would
get applied to all records.

How this addresses that need:
* A new run_timestamp CLI argument
* Passed along to Transmofer .load() and init
* Defaults to run_date, which is required in the input filename,
if not provided

Side effects of this change:
* Transmogrifier will now generate records with a run_timestamp
for writing, allowing multiple Transmogrifier runs to write the
same run_timestamp where applicable

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-509