Add MITLibWebsite transform
Why these changes are being introduced:
This work is to support the addition of MIT Libraries
websites as a new TIMDEX datsa source. The browsertrix-harvester
will be used to generate metadata for live and accessible
websites at the 'libraries.mit.edu' domain.
For these to be added as a new TIMDEX data source,
these records will need to be transformed into
TIMDEX records just like any other source.

How this addresses that need:
* Add new module for MITLibWebsite JSONTransformer
* Add unit tests for MITLibWebsite transforms
* Update dependencies to resolve security vulnerability (authlib)

Relevant ticket(s):
https://mitlibraries.atlassian.net/browse/TIMX-254