Handle deleted records for mitlibwebsite
Why these changes are being introduced:

For the v1 version of the transformation, metadatda records coming
from browsertrix-harvester did not indicate if the record was
deleted, because *all* records were new or modified.

Now that the harvester includes a field called `status` with values
`active` or `deleted`, we can key off this in Transmogrifier to
detect deleted records.

How this addresses that need:

Update the abstract method `record_is_deleted` to look for
`status="deleted"` in the incoming metadata record.

Side effects of this change:
* Tranmogrifier will write deletion records for mitlibwebsite

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/USE-98